+++ packages/angular/test/base/src/app/standalone/standalone.component.ts
@@ -1,8 +1,0 @@
-import { Component } from "@angular/core";
-
-@Component({
-  selector: 'app-standalone',
-  templateUrl: './standalone.component.html',
-  standalone: true
-})
-export class StandaloneComponent { }