+++ packages/vue/scripts/build-web-types.js
@@ -74,13 +74,10 @@ for (const component of filteredComponents) {
     name: componentName,
     "doc-url": docUrl,
     description: component.docs,
+    priority: "highest", // Fixes attribute autocomplete (FW-7636)
     source: {
-      module:
-        "@ionic/core/" +
-        component.filePath
-          .replace("./src/", "dist/types/")
-          .replace(".tsx", ".d.ts"),
-      symbol: componentName.substr(3),
+      module: "@ionic/vue",
+      symbol: componentName,
     },
     attributes,
     slots,